+++ forgerock-auth/src/main/java/org/forgerock/android/auth/StringUtils.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020 ForgeRock. All rights reserved.
+ * Copyright (c) 2020 - 2021 ForgeRock. All rights reserved.
  *
  * This software may be modified and distributed under the terms
  * of the MIT license. See the LICENSE file for details.
@@ -10,7 +10,7 @@ package org.forgerock.android.auth;
 /**
  * Utility class for handling Strings.
  */
-class StringUtils {
+public class StringUtils {
 
     private StringUtils() {
     }