+++ forgerock-auth/src/test/java/org/forgerock/android/auth/SocialLoginTest.java
@@ -11,6 +11,8 @@ package org.forgerock.android.auth;
 import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
+import android.content.pm.PackageManager;
+import android.net.Uri;
 
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentActivity;
@@ -19,11 +21,14 @@ import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
 import com.google.android.gms.common.api.ApiException;
 import com.squareup.okhttp.mockwebserver.RecordedRequest;
 
+import org.forgerock.android.auth.callback.CallbackFactory;
 import org.forgerock.android.auth.callback.IdPCallback;
 import org.forgerock.android.auth.callback.SelectIdPCallback;
+import org.forgerock.android.auth.idp.AppleSignInHandler;
 import org.forgerock.android.auth.idp.GoogleSignInHandler;
 import org.json.JSONException;
 import org.json.JSONObject;
+import org.junit.After;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.Robolectric;
@@ -50,6 +55,20 @@ public class SocialLoginTest extends BaseTest {
             return null;
         }
         return (GoogleSignInHandler) fragment;
+    }
+
+    private AppleSignInHandler getAppleSignInHandler(FragmentActivity activity) {
+        Fragment fragment = activity.getSupportFragmentManager()
+                .findFragmentByTag(AppleSignInHandler.TAG);
+        if (fragment == null) {
+            return null;
+        }
+        return (AppleSignInHandler) fragment;
+    }
+
+    @After
+    public void resetIdPCallback() {
+        CallbackFactory.getInstance().register(IdPCallback.class);
     }
 
     @Test
@@ -257,6 +276,168 @@ public class SocialLoginTest extends BaseTest {
         getGoogleSignInHandler(fragmentActivity).onActivityResult(GoogleSignInHandler.RC_SIGN_IN, Activity.RESULT_OK, null);
         countDownLatch.await();
     }
+
+    @Test
+    public void testHappyPathWithApple() throws InterruptedException, ExecutionException, JSONException, PackageManager.NameNotFoundException {
+
+        CallbackFactory.getInstance().register(IdPCallbackMock.class);
+
+        enqueue("/authTreeMockTest_Authenticate_SelectIdPCallback.json", HttpURLConnection.HTTP_OK);
+        enqueue("/authTreeMockTest_Authenticate_IdPCallback_apple.json", HttpURLConnection.HTTP_OK);
+        enqueue("/authTreeMockTest_Authenticate_success.json", HttpURLConnection.HTTP_OK);
+
+        Config.getInstance().setSharedPreferences(context.getSharedPreferences(DEFAULT_TOKEN_MANAGER_TEST, Context.MODE_PRIVATE));
+        Config.getInstance().setSsoSharedPreferences(context.getSharedPreferences(DEFAULT_SSO_TOKEN_MANAGER_TEST, Context.MODE_PRIVATE));
+        Config.getInstance().setUrl(getUrl());
+        Config.getInstance().setEncryptor(new MockEncryptor());
+
+        FragmentActivity fragmentActivity = Robolectric.buildActivity(FragmentActivity.class).setup().get();
+        InitProvider.setCurrentActivity(fragmentActivity);
+
+        CountDownLatch executeTree = new CountDownLatch(2);
+
+        final IdPCallbackMock[] idPCallback = {null};
+        final Node[] finalState = {null};
+        NodeListenerFuture<FRSession> nodeListenerFuture = new NodeListenerFuture<FRSession>() {
+
+            @Override
+            public void onCallbackReceived(Node state) {
+                if (state.getCallback(SelectIdPCallback.class) != null) {
+                    state.getCallback(SelectIdPCallback.class).setValue("apple");
+                    state.next(context, this);
+                    executeTree.countDown();
+                    return;
+                }
+
+                if (state.getCallback(IdPCallbackMock.class) != null) {
+                    //idPCallback.signIn needs to be run in Main thread in order to
+                    //launch the Fragment
+                    idPCallback[0] = state.getCallback(IdPCallbackMock.class);
+                    finalState[0] = state;
+                    executeTree.countDown();
+                }
+            }
+        };
+
+        FRSession.authenticate(context, "", nodeListenerFuture);
+
+        //Wait for the idPCallback to finish
+        executeTree.await();
+
+        idPCallback[0].signIn(new FRListener<Void>() {
+            @Override
+            public void onSuccess(Void result) {
+                finalState[0].next(context, nodeListenerFuture);
+            }
+
+            @Override
+            public void onException(Exception e) {
+                fail(e.getMessage());
+            }
+        });
+
+        Intent intent = new Intent();
+        intent.setData(Uri.parse("https://opeam.example.com?form_post_entry=dummyValue"));
+        getAppleSignInHandler(fragmentActivity).onActivityResult(GoogleSignInHandler.RC_SIGN_IN, Activity.RESULT_OK, intent);
+
+        assertNotNull(nodeListenerFuture.get());
+        assertNotNull(FRUser.getCurrentUser());
+
+        RecordedRequest recordedRequest = server.takeRequest(); //First request
+        //SelectIdPCallback
+        recordedRequest = server.takeRequest();
+        String body = recordedRequest.getBody().readUtf8();
+        JSONObject result = new JSONObject(body);
+        assertThat(result.getJSONArray("callbacks").length()).isEqualTo(1);
+        JSONObject selectIdPCallback = result.getJSONArray("callbacks").getJSONObject(0);
+        assertThat(selectIdPCallback.getString("type")).isEqualTo("SelectIdPCallback");
+        assertThat(selectIdPCallback.getJSONArray("input").length()).isEqualTo(1);
+        String value = selectIdPCallback.getJSONArray("input").getJSONObject(0).getString("value");
+        assertThat(value).isEqualTo("apple");
+
+        //IdPCallback
+        recordedRequest = server.takeRequest(); //Select IdPCallback
+        body = recordedRequest.getBody().readUtf8();
+        result = new JSONObject(body);
+        assertThat(result.getJSONArray("callbacks").length()).isEqualTo(1);
+        JSONObject idPCallbackReq = result.getJSONArray("callbacks").getJSONObject(0);
+        assertThat(idPCallbackReq.getString("type")).isEqualTo("IdPCallback");
+        assertThat(idPCallbackReq.getJSONArray("input").length()).isEqualTo(2);
+        String token = idPCallbackReq.getJSONArray("input").getJSONObject(0).getString("value");
+        assertThat(token).isEqualTo("form_post_entry");
+        String tokenType = idPCallbackReq.getJSONArray("input").getJSONObject(1).getString("value");
+        assertThat(tokenType).isEqualTo("authorization_code");
+        assertThat(Uri.parse(recordedRequest.getPath()).getQueryParameter("form_post_entry")).isEqualTo("dummyValue");
+
+    }
+
+    @Test
+    public void testAppleSignInWithoutNonce() throws InterruptedException, ExecutionException, JSONException {
+
+        CallbackFactory.getInstance().register(IdPCallbackMock.class);
+
+        enqueue("/authTreeMockTest_Authenticate_SelectIdPCallback.json", HttpURLConnection.HTTP_OK);
+        enqueue("/authTreeMockTest_Authenticate_IdPCallback_apple_no_nonce.json", HttpURLConnection.HTTP_OK);
+        enqueue("/authTreeMockTest_Authenticate_success.json", HttpURLConnection.HTTP_OK);
+
+        Config.getInstance().setSharedPreferences(context.getSharedPreferences(DEFAULT_TOKEN_MANAGER_TEST, Context.MODE_PRIVATE));
+        Config.getInstance().setSsoSharedPreferences(context.getSharedPreferences(DEFAULT_SSO_TOKEN_MANAGER_TEST, Context.MODE_PRIVATE));
+        Config.getInstance().setUrl(getUrl());
+        Config.getInstance().setEncryptor(new MockEncryptor());
+
+        FragmentActivity fragmentActivity = Robolectric.buildActivity(FragmentActivity.class).setup().get();
+        InitProvider.setCurrentActivity(fragmentActivity);
+
+        CountDownLatch executeTree = new CountDownLatch(2);
+
+        final IdPCallbackMock[] idPCallback = {null};
+        final Node[] finalState = {null};
+        NodeListenerFuture<FRSession> nodeListenerFuture = new NodeListenerFuture<FRSession>() {
+
+            @Override
+            public void onCallbackReceived(Node state) {
+                if (state.getCallback(SelectIdPCallback.class) != null) {
+                    state.getCallback(SelectIdPCallback.class).setValue("apple");
+                    state.next(context, this);
+                    executeTree.countDown();
+                    return;
+                }
+
+                if (state.getCallback(IdPCallbackMock.class) != null) {
+                    //idPCallback.signIn needs to be run in Main thread in order to
+                    //launch the Fragment
+                    idPCallback[0] = state.getCallback(IdPCallbackMock.class);
+                    finalState[0] = state;
+                    executeTree.countDown();
+                }
+            }
+        };
+
+        FRSession.authenticate(context, "", nodeListenerFuture);
+
+        //Wait for the idPCallback to finish
+        executeTree.await();
+
+        final boolean[] result = {false};
+        CountDownLatch signInCountDownLatch = new CountDownLatch(1);
+        idPCallback[0].signIn(new FRListener<Void>() {
+            @Override
+            public void onSuccess(Void result) {
+                fail("Expect fail when no nonce");
+            }
+
+            @Override
+            public void onException(Exception e) {
+                result[0] = true;
+                assertThat(e).isInstanceOf(IllegalArgumentException.class);
+                signInCountDownLatch.countDown();
+            }
+        });
+
+        signInCountDownLatch.await();
+        assertThat(result[0]).isTrue();
+    }
+
 
     @Test
     public void testUnsupportedHandler() throws InterruptedException {